fixing select component
